move plotting from utils to analysis
